ADD security with token

+++ src/routes/article.js
@@ -1,8 +1,9 @@
 const router = require('express').Router();
 const axios = require('axios');
-const jwt = require('jsonwebtoken');
+const { checkTokenMiddleware } = require('../middleware/auth');
 const handlerArticle = require('../utils/handler.Article');
 const handlerRestaurant = require('../utils/handler.Restaurant');
+const tokenReceiver = require('../utils/tokenReceiver');
 
 // GET /api/v1/article
 router.get('/', function(req, res) {
@@ -38,16 +39,22 @@ router.get('/restaurant/:id', function(req, res) {
 });
 
 // POST /api/v1/article
-router.post('/', function(req, res) {
-    axios.post(`${handlerArticle()}`, req.body).then(function(response){
-        res.json(response.data);
-        updateRestaurant(response.data.restaurantId, response.data._id);
-        return response.data;
-    }).catch(function(err){
-        console.log(err);
-        res.status(500).json({ message: 'Error. Internal server error' });
-    });
+router.post('/', checkTokenMiddleware, function(req, res) {
 
+    const tokenLoad = extractBearerToken(req);
+
+    if(tokenLoad.role === 'role_technique' || tokenLoad.role === 'role_restaurateur' || tokenLoad.role === 'role_commercial'){
+        axios.post(`${handlerArticle()}`, req.body).then(function(response){
+            res.json(response.data);
+            updateRestaurant(response.data.restaurantId, response.data._id);
+            return response.data;
+        }).catch(function(err){
+            console.log(err);
+            res.status(500).json({ message: 'Error. Internal server error' });
+        });
+    }else{
+        res.status(403).json({ message: 'Error. Forbidden' });
+    }
 });
 
 async function updateRestaurant(id, articleId){
@@ -72,27 +79,38 @@ async function updateRestaurant(id, articleId){
 }
 
 // PUT /api/v1/article
-router.put('/:id', function(req, res) {
-    axios.put(`${handlerArticle()}` + req.params.id, req.body).then(function(response){
-        res.json(response.data);
-        return response.data;
-    }).catch(function(err){
-        console.log(err);
-        res.status(500).json({ message: 'Error. Internal server error' });
-    });
+router.put('/:id', checkTokenMiddleware, function(req, res) {
+    const tokenLoad = extractBearerToken(req);
+
+    if(tokenLoad.role === 'role_technique' || tokenLoad.role === 'role_restaurateur' || tokenLoad.role === 'role_commercial'){
+        axios.put(`${handlerArticle()}` + req.params.id, req.body).then(function(response){
+            res.json(response.data);
+            return response.data;
+        }).catch(function(err){
+            console.log(err);
+            res.status(500).json({ message: 'Error. Internal server error' });
+        });
+    }else{
+        res.status(403).json({ message: 'Error. Forbidden' });
+    }
 });
 
 // DELETE /api/v1/article/:id
-router.delete('/:id', function(req, res) {
-    axios.delete(`${handlerArticle()}` + req.params.id).then(function(response){
-        res.json(response.data);
-        updateRestaurantDelete(response.data.restaurantId, response.data._id);
-        return response.data;
-    }).catch(function(err){
-        console.log(err);
-        res.status(500).json({ message: 'Error. Internal server error' });
-    });
+router.delete('/:id', checkTokenMiddleware, function(req, res) {
+    const tokenLoad = tokenReceiver(req);
 
+    if(tokenLoad.role === 'role_technique' || tokenLoad.role === 'role_restaurateur' || tokenLoad.role === 'role_commercial'){
+        axios.delete(`${handlerArticle()}` + req.params.id).then(function(response){
+            res.json(response.data);
+            updateRestaurantDelete(response.data.restaurantId, response.data._id);
+            return response.data;
+        }).catch(function(err){
+            console.log(err);
+            res.status(500).json({ message: 'Error. Internal server error' });
+        });
+    }else{
+        res.status(403).json({ message: 'Error. Forbidden' });
+    }
 });
 
 async function updateRestaurantDelete(id, articleId){
